# openapi.yaml
openapi: 3.0.3
info:
  title: WikiLivee API
  version: 3.0.0
  description: >
    Wiki-редактор с совместным редактированием в реальном времени.
    Страницы, таблицы (локальное хранение в PostgreSQL), backlinks,
    collaboration (WebSocket), версионирование, якорные комментарии,
    авторизация через JWT.
  license:
    name: MIT
servers:
  - url: https://balooai.ru
    description: Production
  - url: http://localhost:8080
    description: Локальная разработка
security:
  - bearerAuth: []
paths:
...
  /api/tables:
    get:
      operationId: listTables
      tags: [Tables]
      summary: Список всех таблиц
      responses:
        '200':
          description: Список таблиц
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createTable
      tags: [Tables]
      summary: Создать таблицу с колонками
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTableRequest'
            example:
              name: Задачи
              columns:
                - name: Название
                  type: text
                - name: Статус
                  type: text
      responses:
        '201':
          description: Таблица создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/tables/{id}:
    get:
      operationId: getTable
      tags: [Tables]
      summary: Получить таблицу со всеми колонками и строками
      parameters:
        - $ref: '#/components/parameters/tableId'
      responses:
        '200':
          description: Таблица
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateTable
      tags: [Tables]
      summary: Переименовать таблицу
      parameters:
        - $ref: '#/components/parameters/tableId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Таблица обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteTable
      tags: [Tables]
      summary: Удалить таблицу (CASCADE удаляет колонки, строки, ячейки)
      parameters:
        - $ref: '#/components/parameters/tableId'
      responses:
        '204':
          description: Удалено
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tables/{id}/columns:
    post:
      operationId: addColumn
      tags: [Tables]
      summary: Добавить колонку в таблицу
      parameters:
        - $ref: '#/components/parameters/tableId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddColumnRequest'
      responses:
        '201':
          description: Колонка добавлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableColumn'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/tables/{id}/columns/{colId}:
    delete:
      operationId: deleteColumn
      tags: [Tables]
      summary: Удалить колонку (CASCADE удаляет ячейки)
      parameters:
        - $ref: '#/components/parameters/tableId'
        - name: colId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID колонки
      responses:
        '204':
          description: Удалено
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/tables/{id}/rows/{rowId}:
    put:
      operationId: updateRow
      tags: [Tables]
      summary: Обновить ячейки строки (upsert). values — map columnId → value
      parameters:
        - $ref: '#/components/parameters/tableId'
        - name: rowId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID строки
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRowRequest'
            example:
              values:
                "col-uuid-1": "Задача А"
                "col-uuid-2": "В работе"
      responses:
        '200':
          description: Строка обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteRow
      tags: [Tables]
      summary: Удалить строку
      parameters:
        - $ref: '#/components/parameters/tableId'
        - name: rowId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Удалено
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════════════════════
  # WEBSOCKET
  # ════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT токен выдаётся при /api/auth/login или /api/auth/register.
        Срок действия — 30 дней.
  parameters:
...
    tableId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID таблицы
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Требуется авторизация (Bearer токен отсутствует или невалиден)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: page not found

    # ── Auth ─────────────────────────────────────────────
...
    TableSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Table:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        columns:
          type: array
          items:
            $ref: '#/components/schemas/TableColumn'
        rows:
          type: array
          items:
            $ref: '#/components/schemas/TableRow'
    CreateTableRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        columns:
          type: array
          items:
            $ref: '#/components/schemas/TableColumnSpec'
          default: []
    TableColumnSpec:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [text, number, date, select]
          default: text
    AddColumnRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [text, number, date, select]
          default: text
    UpdateRowRequest:
      type: object
      required: [values]
      properties:
        values:
          type: object
          additionalProperties:
            type: string
          description: Map columnId (UUID) → string value. Только указанные ячейки обновляются (upsert)
    TableColumn:
      type: object
      properties:
        id:
          type: string
          format: uuid
        table_id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [text, number, date, select]
        position:
          type: integer
          description: Порядковый номер колонки (0-based)
    TableRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        table_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        values:
          type: object
          additionalProperties:
            type: string
          description: Map columnId (UUID) → string value